refactor(core): move all rendering and cleanup logic into core
