+++ .gitignore
@@ -12,7 +12,7 @@ package-lock.json
 yarn.lock
 
 # generated typing output
-types
+dist
 *.tsbuildinfo
 
 # copied documentation
