--- packages/svelte/src/core/validate-options.js
@@ -1,41 +1,0 @@
-import { ALLOWED_MOUNT_OPTIONS } from './mount.js'
-
-class UnknownSvelteOptionsError extends TypeError {
-  constructor(unknownOptions, allowedOptions) {
-    super(`Unknown options.
-
-    Unknown: [ ${unknownOptions.join(', ')} ]
-    Allowed: [ ${allowedOptions.join(', ')} ]
-
-    To pass both Svelte options and props to a component,
-    or to use props that share a name with a Svelte option,
-    you must place all your props under the \`props\` key:
-
-    render(Component, { props: { /** props here **/ } })
-`)
-    this.name = 'UnknownSvelteOptionsError'
-  }
-}
-
-const validateOptions = (options) => {
-  const isProps = !Object.keys(options).some((option) =>
-    ALLOWED_MOUNT_OPTIONS.includes(option)
-  )
-
-  if (isProps) {
-    return { props: options }
-  }
-
-  // Check if any props and Svelte options were accidentally mixed.
-  const unknownOptions = Object.keys(options).filter(
-    (option) => !ALLOWED_MOUNT_OPTIONS.includes(option)
-  )
-
-  if (unknownOptions.length > 0) {
-    throw new UnknownSvelteOptionsError(unknownOptions, ALLOWED_MOUNT_OPTIONS)
-  }
-
-  return options
-}
-
-export { UnknownSvelteOptionsError, validateOptions }
